Target Info Change
casting actors to the right type
methods for changing targets
